Refactor integration tests to adopt other catalogs
