+++ kafka-connect-runtime/src/test/java/io/tabular/iceberg/connect/AbstractIntegrationMultiTableTest.java
@@ -18,9 +18,6 @@
  */
 package io.tabular.iceberg.connect;
 
-import static io.tabular.iceberg.connect.TestConstants.AWS_ACCESS_KEY;
-import static io.tabular.iceberg.connect.TestConstants.AWS_REGION;
-import static io.tabular.iceberg.connect.TestConstants.AWS_SECRET_KEY;
 import static io.tabular.iceberg.connect.TestEvent.TEST_SCHEMA;
 import static io.tabular.iceberg.connect.TestEvent.TEST_SPEC;
 import static java.lang.String.format;
@@ -29,13 +26,10 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.time.Duration;
 import java.util.List;
-import org.apache.iceberg.CatalogProperties;
-import org.apache.iceberg.CatalogUtil;
 import org.apache.iceberg.DataFile;
 import org.apache.iceberg.Table;
-import org.apache.iceberg.aws.AwsClientProperties;
-import org.apache.iceberg.aws.s3.S3FileIOProperties;
 import org.apache.iceberg.catalog.Namespace;
+import org.apache.iceberg.catalog.SupportsNamespaces;
 import org.apache.iceberg.catalog.TableIdentifier;
 import org.awaitility.Awaitility;
 import org.junit.jupiter.api.AfterEach;
@@ -44,55 +38,49 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-public class IntegrationTest extends IntegrationTestBase {
+public abstract class AbstractIntegrationMultiTableTest extends IntegrationTestBase {
 
   private static final String TEST_DB = "test";
-  private static final String TEST_TABLE = "foobar";
-  private static final TableIdentifier TABLE_IDENTIFIER = TableIdentifier.of(TEST_DB, TEST_TABLE);
+  private static final String TEST_TABLE1 = "foobar1";
+  private static final String TEST_TABLE2 = "foobar2";
+  private static final TableIdentifier TABLE_IDENTIFIER1 = TableIdentifier.of(TEST_DB, TEST_TABLE1);
+  private static final TableIdentifier TABLE_IDENTIFIER2 = TableIdentifier.of(TEST_DB, TEST_TABLE2);
 
   @BeforeEach
   public void setup() {
     createTopic(testTopic, TEST_TOPIC_PARTITIONS);
-    catalog.createNamespace(Namespace.of(TEST_DB));
+    ((SupportsNamespaces) catalog).createNamespace(Namespace.of(TEST_DB));
   }
 
   @AfterEach
   public void teardown() {
-    context.stopConnector(connectorName);
+    context.stopKafkaConnector(connectorName);
     deleteTopic(testTopic);
-    catalog.dropTable(TableIdentifier.of(TEST_DB, TEST_TABLE));
-    catalog.dropNamespace(Namespace.of(TEST_DB));
+    catalog.dropTable(TableIdentifier.of(TEST_DB, TEST_TABLE1));
+    catalog.dropTable(TableIdentifier.of(TEST_DB, TEST_TABLE2));
+    ((SupportsNamespaces) catalog).dropNamespace(Namespace.of(TEST_DB));
   }
 
   @ParameterizedTest
   @NullSource
   @ValueSource(strings = {"test_branch"})
-  public void testIcebergSinkPartitionedTable(String branch) {
-    catalog.createTable(TABLE_IDENTIFIER, TEST_SCHEMA, TEST_SPEC);
+  public void testIcebergSink(String branch) {
+    // partitioned table
+    catalog.createTable(TABLE_IDENTIFIER1, TEST_SCHEMA, TEST_SPEC);
+    // unpartitioned table
+    catalog.createTable(TABLE_IDENTIFIER2, TEST_SCHEMA);
 
     runTest(branch);
 
-    List<DataFile> files = getDataFiles(TABLE_IDENTIFIER, branch);
-    // partition may involve 1 or 2 workers
-    assertThat(files).hasSizeBetween(1, 2);
+    List<DataFile> files = getDataFiles(TABLE_IDENTIFIER1, branch);
+    assertThat(files).hasSize(1);
     assertEquals(1, files.get(0).recordCount());
-    assertEquals(1, files.get(1).recordCount());
-    assertSnapshotProps(TABLE_IDENTIFIER, branch);
-  }
+    assertSnapshotProps(TABLE_IDENTIFIER1, branch);
 
-  @ParameterizedTest
-  @NullSource
-  @ValueSource(strings = {"test_branch"})
-  public void testIcebergSinkUnpartitionedTable(String branch) {
-    catalog.createTable(TABLE_IDENTIFIER, TEST_SCHEMA);
-
-    runTest(branch);
-
-    List<DataFile> files = getDataFiles(TABLE_IDENTIFIER, branch);
-    // may involve 1 or 2 workers
-    assertThat(files).hasSizeBetween(1, 2);
-    assertEquals(2, files.stream().mapToLong(DataFile::recordCount).sum());
-    assertSnapshotProps(TABLE_IDENTIFIER, branch);
+    files = getDataFiles(TABLE_IDENTIFIER2, branch);
+    assertThat(files).hasSize(1);
+    assertEquals(1, files.get(0).recordCount());
+    assertSnapshotProps(TABLE_IDENTIFIER2, branch);
   }
 
   private void runTest(String branch) {
@@ -107,32 +95,30 @@ public class IntegrationTest extends IntegrationTestBase {
             .config("key.converter.schemas.enable", false)
             .config("value.converter", "org.apache.kafka.connect.json.JsonConverter")
             .config("value.converter.schemas.enable", false)
-            .config("iceberg.tables", format("%s.%s", TEST_DB, TEST_TABLE))
+            .config(
+                "iceberg.tables",
+                format("%s.%s, %s.%s", TEST_DB, TEST_TABLE1, TEST_DB, TEST_TABLE2))
+            .config("iceberg.tables.routeField", "type")
+            .config(format("iceberg.table.%s.%s.routeRegex", TEST_DB, TEST_TABLE1), "type1")
+            .config(format("iceberg.table.%s.%s.routeRegex", TEST_DB, TEST_TABLE2), "type2")
             .config("iceberg.control.commitIntervalMs", 1000)
             .config("iceberg.control.commitTimeoutMs", Integer.MAX_VALUE)
-            .config(
-                "iceberg.catalog." + CatalogUtil.ICEBERG_CATALOG_TYPE,
-                CatalogUtil.ICEBERG_CATALOG_TYPE_REST)
-            .config("iceberg.catalog." + CatalogProperties.URI, "http://iceberg:8181")
-            .config("iceberg.catalog." + S3FileIOProperties.ENDPOINT, "http://minio:9000")
-            .config("iceberg.catalog." + S3FileIOProperties.ACCESS_KEY_ID, AWS_ACCESS_KEY)
-            .config("iceberg.catalog." + S3FileIOProperties.SECRET_ACCESS_KEY, AWS_SECRET_KEY)
-            .config("iceberg.catalog." + S3FileIOProperties.PATH_STYLE_ACCESS, true)
-            .config("iceberg.catalog." + AwsClientProperties.CLIENT_REGION, AWS_REGION)
             .config("iceberg.kafka.auto.offset.reset", "earliest");
+    connectorCatalogProperties().forEach(connectorConfig::config);
 
     if (branch != null) {
       connectorConfig.config("iceberg.tables.defaultCommitBranch", branch);
     }
-    context.startConnector(connectorConfig);
+
+    context.startKafkaConnector(connectorConfig);
 
     TestEvent event1 = new TestEvent(1, "type1", System.currentTimeMillis(), "hello world!");
-
-    long threeDaysAgo = System.currentTimeMillis() - Duration.ofDays(3).toMillis();
-    TestEvent event2 = new TestEvent(2, "type2", threeDaysAgo, "having fun?");
+    TestEvent event2 = new TestEvent(2, "type2", System.currentTimeMillis(), "having fun?");
+    TestEvent event3 = new TestEvent(3, "type3", System.currentTimeMillis(), "ignore me");
 
     send(testTopic, event1);
     send(testTopic, event2);
+    send(testTopic, event3);
     flush();
 
     Awaitility.await()
@@ -142,7 +128,9 @@ public class IntegrationTest extends IntegrationTestBase {
   }
 
   private void assertSnapshotAdded() {
-    Table table = catalog.loadTable(TABLE_IDENTIFIER);
+    Table table = catalog.loadTable(TABLE_IDENTIFIER1);
+    assertThat(table.snapshots()).hasSize(1);
+    table = catalog.loadTable(TABLE_IDENTIFIER2);
     assertThat(table.snapshots()).hasSize(1);
   }
 }
